Initial implementation of new new home screen
