+++ app/src/main/AndroidManifest.xml
@@ -14,10 +14,13 @@
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-    </application>
+        <activity
+            android:name=".ui.CollectionActivity"
+            android:exported="true">
 
+        </activity>
+    </application>
 </manifest>
