+++ app/src/main/java/com/jamieco/movieapp/data/MovieCollection.kt
@@ -1,0 +1,5 @@
+package com.jamieco.movieapp.data
+
+data class MovieCollection(
+    val collectionId: String
+)
