+++ app/src/main/java/com/jamieco/movieapp/data/ResponseObject.kt
@@ -1,4 +1,4 @@
-package com.jamieco.movieapp
+package com.jamieco.movieapp.data
 
 data class ResponseObject(
     val page: Int
