+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".ui.HomeActivity">
+        tools:context=".ui.CollectionActivity">
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content">
